+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+- added lazily injecting the default esri css style
 ### Changed
 - added GitHub issue and pull request templates
 - added badges to README